requirements.txt
bitsandbytes==0.43.3
datasets==3.6.0
fsspec==2023.12.2
loralib==0.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
peft==0.11.1
sympy==1.14.0
tokenizers==0.20.3
torch==2.3.0
transformers==4.45.0
triton==2.3.0
xformers==0.0.26.post1
huggingface-hub==0.33.0
pillow==11.2.1